% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_countries_arg.R
\name{check_cols_exists}
\alias{check_cols_exists}
\title{Check that the countries object has the specified columns}
\usage{
check_cols_exists(countries_object, cols)
}
\arguments{
\item{countries_object}{\code{countries} object to check}

\item{cols}{columns to ensure exist in \code{countries_object}}
}
\description{
Check that the countries object has the specified columns
}
