% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_countries.R
\name{compare_countries}
\alias{compare_countries}
\title{Use the \code{daff} R package to compare versions of the countries object}
\usage{
compare_countries(
  new,
  old,
  col_order = c("iso3", "who_member", "who_short_name_en")
)
}
\arguments{
\item{new}{[\code{data.frame(1)}]\cr
The new data frame to check for new changes compared to the old data frame.}

\item{old}{[\code{data.frame(1)}]\cr
The reference (old) data frame to compare changes against.}

\item{col_order}{[\code{character()}]\cr
Column order for the returned workbook object.}
}
\value{
a named [\code{list(2)}] with 'daff' storing the difference object and
'wb_daff' storing the \code{openxlsx2::wbWorkbook} object.
}
\description{
Use the \code{daff} R package to compare versions of the countries object
}
\examples{
\dontrun{
countries_new_xmart <- make_countries(
  wb_use_xmart = TRUE,
  unsd_special_use_xmart = TRUE,
  sdg_use_xmart = TRUE
)
diff_against_previous <- compare_countries(
  new = countries_new_xmart,
  old = countries
)
daff::render_diff(
  diff = diff_against_previous$daff,
  title = "Comparison of previous and new `whoville::countries` object"
)
openxlsx2::wb_save(
  wb = diff_against_previous$wb_daff,
  file = stringr::str_glue("{here::here()}/data-raw/countries_diff_against_previous.xlsx")
)
}
}
