% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{country_code_types}
\alias{country_code_types}
\title{Country code types available}
\usage{
country_code_types(countries_manual = NULL)
}
\arguments{
\item{countries_manual}{Optional argument to directly pass in the countries
object to use. If specified the object is used, otherwise the
'whoville.updated.countries' option is used if it is set, and finally
otherwise the built in \code{whoville::countries} object is used.}
}
\value{
Character vector of country code types.
}
\description{
\code{country_code_types()} provides a character vector of all country code types
available in the \code{countries} data frame.
}
\examples{
country_code_types()
}
