% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_who_xmart.R
\name{format_who_xmart_ref_country}
\alias{format_who_xmart_ref_country}
\title{Format the country list from the WHO xMart REFMART/REF_COUNTRY table}
\usage{
format_who_xmart_ref_country(dat_who_xmart_ref_country)
}
\arguments{
\item{dat_who_xmart_ref_country}{[\code{tibble(1)}]\cr
data returned from \code{get_who_public_xmart}}
}
\value{
[\code{tibble()}] with formatted WHO country list with data for the
\code{whoville::countries} object.
}
\description{
Format the country list from the WHO xMart REFMART/REF_COUNTRY table for
inclusion in the \code{whoville::countries} object
}
\seealso{
\code{\link[=get_who_public_xmart]{get_who_public_xmart()}}
}
