% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/member_states.R
\name{gbd_high_income_states}
\alias{gbd_high_income_states}
\title{Get ISO3 codes for GBD high-income states.}
\usage{
gbd_high_income_states(countries_manual = NULL)
}
\arguments{
\item{countries_manual}{Optional argument to directly pass in the countries
object to use. If specified the object is used, otherwise the
'whoville.updated.countries' option is used if it is set, and finally
otherwise the built in \code{whoville::countries} object is used.}
}
\value{
A character vector of ISO3 codes.
}
\description{
\code{gbd_high_income_states()} returns ISO3 codes for states classified as high
income in the 2019 Global Burden of Disease released by the Institute for
Health Metrics and Evaluation. Useful to expand data frames to explicitly
include missing data for countries.
}
