% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oecd.R
\name{get_oecd_countries_direct}
\alias{get_oecd_countries_direct}
\title{Get Organisation for Economic Co-operation and Development (OECD)
member countries}
\usage{
get_oecd_countries_direct(
  url =
    "https://sdmx.oecd.org/public/rest/data/OECD.SDD.TPS,DSD_CPI_COU_WEIGHTS@DF_CPI_CTRY_WEIGHTS,1.0/.A..?dimensionAtObservation=AllDimensions&format=csvfilewithlabels"
)
}
\arguments{
\item{url}{[\code{character(1)}]\cr
URL to an OECD data-explorer dataset for OECD only countries (currently
the OECD consumer price index country weights)}
}
\value{
[\code{data.frame()}] with a row for each 'iso3' in \code{whoville::countries} and
a logical column called 'oecd_member' indicating membership in OECD.
}
\description{
Get Organisation for Economic Co-operation and Development (OECD)
member countries
}
\examples{
\dontrun{
oecd <- get_oecd_countries_direct()
}
}
