% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_who_xmart.R
\name{get_who_public_xmart}
\alias{get_who_public_xmart}
\title{Get data from a public WHO xMart table.}
\usage{
get_who_public_xmart(
  url = "https://xmart-api-public.who.int/REFMART/REF_COUNTRY"
)
}
\arguments{
\item{url}{[\code{character(1)}]\cr
Url to xMart api, see the xmart documentation for how to modify the url to
filter rows and select columns.}
}
\value{
[\code{tibble()}] with data from the specified mart and table.
}
\description{
Get data from a public WHO xMart table.
}
\details{
This helper function is only set up to get publicly available data from xMart.
If you need more general access, see the \href{https://github.com/gpw13/xmart4}{\code{xmart4}}  R package.
\itemize{
\item \href{https://extranet.who.int/xmart4/docs/}{xMart Documentation}
\item \href{https://extranet.who.int/xmart4/data}{See all available marts}
\item \href{https://extranet.who.int/xmart4/REFMART}{See example REFMART}:
(Reference database of WHO. For example, standard country and area lists, country population estimates, dimensions, etc.)
\item \href{https://extranet.who.int/xmart4/REFMART}{See example REFMART/REF_COUNTRY}:
(COMPREHENSIVE COUNTRY/AREA LIST)
}
}
\examples{
\dontrun{
url <- "https://xmart-api-public.who.int/REFMART/REF_COUNTRY"
get_who_public_xmart(url = url)
url <- "https://xmart-api-public.who.int/DATA_/REF_SEX"
url <- paste0(url, "?$select=REF_TERM,TERM_NAME & $filter=REF_TERM in ('MALE','FEMALE')")
get_who_public_xmart(url = url)
}

}
\seealso{
\code{\link[=format_who_xmart_ref_country]{format_who_xmart_ref_country()}}

\code{\link[xmart4:xmart4-package]{xmart4::xmart4()}}
}
