% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_countries.R
\name{highlight_rows_columns}
\alias{highlight_rows_columns}
\title{Helper function to convert the difference object to an
openxlsx2::wbWorkbook object with cell highlighting and other light formatting.}
\usage{
highlight_rows_columns(
  wb,
  dat,
  dimension = c("rows", "columns"),
  change_type = c("added", "removed")
)
}
\arguments{
\item{wb}{[\code{wbWorkbook(1)}]\cr
\code{openxlsx2} workbook to format with highlighted cells, rows and columns.}

\item{dat}{[\code{data.frame(1)}]\cr
differences data frame returned by the \code{get_data} method from the returned
differences object.}

\item{dimension}{[\code{character(1)}]\cr
whether to highlight 'rows' or 'columns'.}

\item{change_type}{[\code{character(1)}]\cr
whether to highlight \code{dimensions} that were 'added' or 'removed'.}
}
\description{
Helper function to convert the difference object to an
openxlsx2::wbWorkbook object with cell highlighting and other light formatting.
}
