% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/member_states.R
\name{is_gbd_high_income}
\alias{is_gbd_high_income}
\title{Check GBD high-income classification status from ISO3 codes.}
\usage{
is_gbd_high_income(iso3, countries_manual = NULL)
}
\arguments{
\item{iso3}{Character vector of ISO3 codes.}

\item{countries_manual}{Optional argument to directly pass in the countries
object to use. If specified the object is used, otherwise the
'whoville.updated.countries' option is used if it is set, and finally
otherwise the built in \code{whoville::countries} object is used.}
}
\value{
Logical vector.
}
\description{
\code{is_gbd_high_income()} takes in a vector of ISO3 codes and returns a logical vector
on whether that country is classified as high-income in the 2019 Global Burden
of Disease released by the Institute for Health Metrics and Evaluation.
}
