% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/member_states.R
\name{is_oecd_member}
\alias{is_oecd_member}
\title{Get OECD membership status from ISO3 codes.}
\usage{
is_oecd_member(iso3, countries_manual = NULL)
}
\arguments{
\item{iso3}{Character vector of ISO3 codes.}

\item{countries_manual}{Optional argument to directly pass in the countries
object to use. If specified the object is used, otherwise the
'whoville.updated.countries' option is used if it is set, and finally
otherwise the built in \code{whoville::countries} object is used.}
}
\value{
Logical vector.
}
\description{
\code{is_oecd_member()} takes in a vector of ISO3 codes and returns a logical vector
on whether that country is an OECD member state or not.
}
