% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/member_states.R
\name{is_who_member_large}
\alias{is_who_member_large}
\title{Get large state membership status from ISO3 codes.}
\usage{
is_who_member_large(iso3, countries_manual = NULL)
}
\arguments{
\item{iso3}{Character vector of ISO3 codes.}

\item{countries_manual}{Optional argument to directly pass in the countries
object to use. If specified the object is used, otherwise the
'whoville.updated.countries' option is used if it is set, and finally
otherwise the built in \code{whoville::countries} object is used.}
}
\value{
Logical vector.
}
\description{
\code{is_who_member_large()} takes in a vector of ISO3 codes and returns a logical vector
on whether that country is a large WHO member state or not. In this instance,
large member states are defined as member states with a total population >
90,000 in mid-year 2018, according to the World Population Prospects.
}
