% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{iso3_to_regions}
\alias{iso3_to_regions}
\title{Get regions from ISO3 country codes.}
\usage{
iso3_to_regions(
  iso3,
  region = c("who_region", "un_region", "un_subregion", "un_intermediate_region",
    "sdg_region", "sdg_subregion", "gbd_region", "gbd_subregion", "un_desa_region",
    "un_desa_subregion", "wb_region", "wb_ig"),
  year = max(wb_ig_years()),
  name = FALSE,
  language = c("en", "es", "ru", "ar", "zh", "fr"),
  countries_manual = NULL
)
}
\arguments{
\item{iso3}{Character vector of ISO3 codes.}

\item{region}{Type of region to return. Regional classifications available from
the WHO, UN, UNSD SDGs, UN DESA, the World Bank (geographic and income based),
and IHME's GBD. See the \link{countries} documentation for more source
details.}

\item{year}{For World Bank Income Group only, specify the year to return.}

\item{name}{For UN, UNSD SDG, UN DESA, WB and IHME GBD regions only. If \code{TRUE},
return official regional name instead of code.}

\item{language}{For \code{"un_region"}, \code{"un_subregion"}, and \code{"un_intermediate_region"}
only if returning name. A character value specifying the language of the
country names. Should be specified using the ISO2 language code.}

\item{countries_manual}{Optional argument to directly pass in the countries
object to use. If specified the object is used, otherwise the
'whoville.updated.countries' option is used if it is set, and finally
otherwise the built in \code{whoville::countries} object is used.}
}
\value{
Character vector.
}
\description{
\code{iso3_to_regions()} takes in a vector of ISO3 codes and returns a vector of
specified regions. Currently, this can be codes or English names for
UN regions/subregions/intermediate regions, UN DESA regions/subregions,
SDG regions/subregions, WHO regions, and World Bank geographical regions and
Income Groups.
}
