% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_oecd.R
\name{get_oecd_countries}
\alias{get_oecd_countries}
\alias{format_oecd_countries}
\title{Get/Format Organisation for Economic Co-operation and Development (OECD)
member countries}
\usage{
get_oecd_countries(url = "https://www.oecd.org/en/about/members-partners.html")

format_oecd_countries(
  oecd_country_names,
  map_iso3_manual = data.frame(oecd_country_name = c("Korea", "Slovak Republic",
    "Netherlands"), iso3 = c("KOR", "SVK", "NLD")),
  allow_na = FALSE
)
}
\arguments{
\item{url}{[\code{character(1)}]\cr
OECD member country webpage url}

\item{oecd_country_names}{[\code{character()}]\cr Scraped OECD member country names}

\item{map_iso3_manual}{[\code{data.frame()}]\cr
Mapping between \code{oecd_country_name} and \code{iso3} for OECD country names that
do not exactly match one of the \code{whoville::countries} name columns.}

\item{allow_na}{[\code{logical(1)}]\cr
Whether to allow NA values in the result. Switching this to \code{TRUE} can be
helpful to figure out which \code{oecd_country_names} are not automatically
mapping to a \code{iso3}.}
}
\value{
\code{get_oecd_countries()} returns a [\code{character()}] vector with member country
names as presented on the \code{url}. \code{format_oecd_countries()} returns a [\code{data.frame()}]
with a row for each \code{oecd_country_names} and the mapped \code{iso3} code.
}
\description{
Get/Format Organisation for Economic Co-operation and Development (OECD)
member countries
}
\examples{
\dontrun{
oecd <- get_oecd_countries() \%>\%
  format_oecd_countries()
}

}
