% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resolve_countries_arg.R
\name{resolve_countries_arg}
\alias{resolve_countries_arg}
\title{Helper function to resolve the countries object used in basic whoville
functions}
\usage{
resolve_countries_arg(countries_manual)
}
\arguments{
\item{countries_manual}{Optional argument to directly pass in the countries
object to use. If specified the object is used, otherwise the
'whoville.updated.countries' option is used if it is set, and finally
otherwise the built in \code{whoville::countries} object is used.}
}
\description{
Helper function to resolve the countries object used in basic whoville
functions
}
\examples{
\dontrun{
countries_updated <- make_countries()
options(whoville.option.countries = countries_updated)
iso3_to_regions(iso3 = 'JPN', region = 'who_region')
}
}
