% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_un_m49.R
\name{get_un_m49}
\alias{get_un_m49}
\alias{format_un_m49}
\title{Get/Format the United Nations Statistics Division (UNSD) m49 regions
and location name translations.}
\usage{
get_un_m49(url = "https://unstats.un.org/unsd/methodology/m49/overview/")

format_un_m49(
  dat_unsd_m49,
  language_codes = c(ru = "russian", fr = "french", es = "spanish", ar = "arabic", zh =
    "chinese")
)
}
\arguments{
\item{url}{[\code{character(1)}]\cr
Url to scrape the UNSD table (in html format) from.}

\item{dat_unsd_m49}{[\code{data.frame(1)}]\cr
The returned object from \code{get_un_m49} to be formatted to match the
\code{whoville::countries} object formatting.}

\item{language_codes}{[\code{character()}]\cr
Named character vector with the values corresponding to the language names
used in \code{dat_unsd_m49} and names corresponding to the language code to use
in the returned formatted data frame.}
}
\value{
\code{get_un_m49} returns a named list of [\code{data.frame()}]s with the list names
corresponding to the labeled language. \code{format_un_m49} returns a [\code{tibble()}]
with the same data formatted to match the \code{whoville::countries} object.
}
\description{
Get/Format the United Nations Statistics Division (UNSD) m49 regions
and location name translations.
}
\examples{
\dontrun{
un_m49 <- get_un_m49() \%>\%
  format_un_m49()
}

}
