% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{un_population}
\alias{un_population}
\title{Country population data from the UN.}
\format{
A data frame with 30200 rows and 5 variables:
\describe{
\item{iso3}{\href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-3}{ISO3 code}}
\item{year}{Year, currently 1950 - 2100}
\item{female}{Total population - female}
\item{male}{Total population - male}
\item{total}{Total population}
}
}
\source{
\href{https://population.un.org/wpp/Download/Standard/Population/}{United Nations, Department of Economic and Social Affairs, Population Division (2019). World Population Prospects 2019, Online Edition. Rev. 1.}
}
\usage{
un_population
}
\description{
A dataset containing ISO3 code, year, and gender disaggregated population
figures for countries, using the most recent United Nations, Department
of Economic and Social Affairs, Population Division estimates.
}
\details{
Population data is based on the most recent estimates from the United Nations,
Department of Economic and Social Affairs, Population Division. For years
2021 - 2100, the population model assumes medium fertility throughout that
time. See the
\href{https://population.un.org/wpp/DefinitionOfProjectionVariants/}{World
Population Prospect's definition of projection variants} for more details
on the assumptions and model used.
}
\keyword{datasets}
