% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_undesa_sdg.R
\name{get_undesa_sdg_direct}
\alias{get_undesa_sdg_direct}
\alias{format_undesa_sdg_direct}
\alias{format_sdg_xmart}
\title{Get/Format the regions defined by the United Nations Department of
Economic and Social Affairs (UN DESA) and the Sustainable Development Goals (SDG)}
\usage{
get_undesa_sdg_direct(
  url_file =
    "https://population.un.org/wpp/assets/Excel\%20Files/4_Metadata/WPP2024_F01_LOCATIONS.xlsx",
  url_wpp = "https://population.un.org/wpp/",
  wpp_year = 2024
)

format_undesa_sdg_direct(dat_undesa_sdg)

format_sdg_xmart(dat_ref_groups_current)
}
\arguments{
\item{url_file}{[\code{character(1)}]\cr
Url to download the 'Locations (XLSX)' file from, the 'DB' sheet is read in.
This link changes for every WPP revision year.}

\item{url_wpp}{[\code{character(1)}]\cr
Url to the WPP home page. This is used to double check whether a more
recent \code{url_file} link is likely available.}

\item{wpp_year}{[\code{numeric(1)}]\cr
The WPP year that should be expected to be found on the WPP home page.
This is used to double check whether a more recent \code{url_file} link is
likely available.}

\item{dat_undesa_sdg}{[\code{data.frame(1)}]\cr
The returned object from \code{get_undesa_sdg} to be formatted to match the
\code{whoville::countries} object formatting.}

\item{dat_ref_groups_current}{[\code{data.frame(1)}]\cr
The xMart4 REFMART 'REF_GROUPS_CURRENT' table to be formatted to match the
\code{whoville::countries} object formatting.}
}
\value{
\code{format_sdg_xmart} returns a [\code{tibble()}] with the SDG regions definied in
xMart 'REFMART/REF_GROUPS_CURRENT' formatted to match the
\code{whoville::countries} object.

\code{get_undesa_sdg_direct} returns a [\code{tibble()}] with the raw data found
in the 'Locations (XLSX)' document 'DB' sheet. \code{format_undesa_sdg_direct} returns a
[\code{tibble()}] with the same data formatted to match the
\code{whoville::countries} object.
}
\description{
Get and format the UN DESA and SDG regions defined in the World Population Prospects (WPP)
locations \href{https://population.un.org/wpp/downloads?folder=Documentation&group=Documentation}{documentation}.
The \href{https://population.un.org/wpp/definition-of-regions}{WPP 'Definition of Regions' page}
also describes this in more detail.
}
\examples{
\dontrun{
dat_sdg <- format_sdg_xmart(
  dat_ref_groups_current =  get_who_public_xmart(
    url = "https://xmart-api-public.who.int/REFMART/REF_GROUPS_CURRENT"
  )
)

undesa_sdg <- get_undesa_sdg_direct() \%>\%
  format_undesa_sdg_direct()
}

}
