% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{valid_iso3}
\alias{valid_iso3}
\title{Check if ISO3 codes present in \code{countries} data frame.}
\usage{
valid_iso3(iso3)
}
\arguments{
\item{iso3}{Character vector of ISO3 codes.}
}
\value{
Logical vector.
}
\description{
\code{valid_iso3()} takes in a vector of ISO3 codes and returns a logical vector
on whether that country code is present in \code{countries[["iso3"]]}
}
