% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wb.R
\name{get_wb_ig}
\alias{get_wb_ig}
\alias{get_wb_reg}
\alias{format_wb_ig}
\alias{format_wb_reg}
\title{Get/Format the World Bank Income Group and Region country classifications.}
\usage{
get_wb_ig(
  url =
    "https://datacatalogapi.worldbank.org/ddhxext/ResourceDownload?resource_unique_id=DR0090754",
  sheet = "Country Analytical History",
  skip_years = 5,
  skip_data = 10
)

get_wb_reg(
  url =
    "https://datacatalogapi.worldbank.org/ddhxext/ResourceDownload?resource_unique_id=DR0090755",
  sheet = "List of economies"
)

format_wb_ig(dat)

format_wb_reg(dat)
}
\arguments{
\item{url}{[\code{character(1)}]\cr
Url to download the country classifications if xlsx format.}

\item{sheet}{[\code{character(1)}]\cr
The name of the sheet within the xlsx document where the country
classifications are stored.}

\item{skip_years}{[\code{integer(1)}]\cr
The number of rows to skip when reading the XLSX document before reaching
the row listing the years available in the historical classification by income group document.}

\item{skip_data}{[\code{integer(1)}]\cr
The number of rows to skip when reading the XLSX document before reaching
the data rows with the country classifications.}

\item{dat}{[\code{data.frame(1)}]\cr
The original country classification data from the World Bank to be formatted
to match the \code{whoville::countries} object formatting.}
}
\value{
\code{get_wb_ig} and \code{get_wb_reg} return a [\code{tibble()}] with data from the
specified XLSX document. \code{format_wb_ig} and \code{format_wb_reg} return a
[\code{tibble()}] with the same data formatted to match the \code{whoville::countries}
object formatting.
}
\description{
Each fiscal year, the World Bank updates their income group country
classifications and releases the files
\href{https://datahelpdesk.worldbank.org/knowledgebase/articles/906519-world-bank-country-and-lending-groups}{here}.
The links to the "historical classification by income in XLSX format" and
the "current classification by income in XLSX format" include the World Bank
Income Group and World Bank Region classifications respectively.
}
\examples{
\dontrun{
wb_ig <- get_wb_ig() \%>\%
  format_wb_ig()
wb_reg <- get_wb_reg() \%>\%
  format_wb_reg()
}

}
