% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{codes_to_iso3}
\alias{codes_to_iso3}
\title{Get ISO3 codes from other country codes.}
\usage{
codes_to_iso3(codes, type)
}
\arguments{
\item{codes}{Character vector of country codes.}

\item{type}{A character value specifying the type of country code supplied.
All possible values available through \code{country_code_types()}.}
}
\value{
Character vector.
}
\description{
\code{codes_to_iso3()} takes in a vector of country codes and returns ISO3 codes.
}
