% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{countries}
\alias{countries}
\title{Country metadata from the WHO, UN, and WB.}
\format{
A data frame with 249 rows and 112 variables:
\describe{
\item{iso3}{\href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-3}{ISO3 code.}}
\item{iso2}{\href{https://en.wikipedia.org/wiki/ISO_3166-1_alpha-2}{ISO2 code.}}
\item{iso_numeric}{\href{https://en.wikipedia.org/wiki/ISO_3166-1_numeric}{ISO numeric code.}}
\item{who_code}{\href{https://apps.who.int/gho/data/node.metadata.COUNTRY?lang=en}{World Health Organization code.}}
\item{m49}{\href{https://en.wikipedia.org/wiki/UN_M49}{United Nations M49 code.}}
\item{gbd_code}{\href{http://www.healthdata.org/gbd/faq}{IHME GBD country code.}}
\item{sovereign_iso3}{ISO3 code of sovereign nation. Value is \code{NA} if sovereign itself.}
\item{who_member}{Logical, is country a WHO member state?}
\item{who_member_small}{Logical, is country a WHO member state with population <= 90,000?}
\item{oecd_member}{Logical, is country an OECD member state?}
\item{un_ldc}{Logical, is the country classified as a least-developed country by the United Nations?}
\item{un_lldc}{Logical, is the country classified as a land-locked developing country by the United Nations?}
\item{un_sids}{Logical, is the country classified as a small island developing state by the United Nations?}
\item{who_short_name_en}{WHO country name, short, English.}
\item{who_formal_name_en}{WHO country name, formal, English. Only available for official member states.}
\item{who_short_name_ar}{WHO country name, short, Arabic.}
\item{who_formal_name_ar}{WHO country name, formal, Arabic. Only available for official member states.}
\item{who_short_name_es}{WHO country name, short, Spanish.}
\item{who_formal_name_es}{WHO country name, formal, Spanish. Only available for official member states.}
\item{who_short_name_fr}{WHO country name, short, French.}
\item{who_formal_name_fr}{WHO country name, formal, French. Only available for official member states.}
\item{who_short_name_ru}{WHO country name, short, Russian.}
\item{who_formal_name_ru}{WHO country name, formal, Russian. Only available for official member states.}
\item{who_short_name_zh}{WHO country name, short, Chinese.}
\item{who_formal_name_zh}{WHO country name, formal, Chinese. Only available for official member states.}
\item{un_name_en}{UN country name, English.}
\item{un_name_ru}{UN country name, Russian.}
\item{un_name_fr}{UN country name, French.}
\item{un_name_es}{UN country name, Spanish.}
\item{un_name_ar}{UN country name, Arabic.}
\item{un_name_zh}{UN country name, Chinese.}
\item{alt_name_en}{Alternative (non-official) country name, English.}
\item{alt_name_2_en}{Another alternative (non-official) country name, English.}
\item{alt_name_3_en}{Another alternative (non-official) country name, English.}
\item{alt_name_4_en}{Another alternative (non-official) country name, English.}
\item{alt_name_5_en}{Another alternative (non-official) country name, English.}
\item{former_name_en}{Former country (non-official) name, English.}
\item{former_name_2_en}{Another former (non-official) country name, English.}
\item{who_region}{WHO region code.}
\item{un_region}{UN region code.}
\item{un_subregion}{UN sub-region code.}
\item{un_intermediate_region}{UN intermediate region code.}
\item{un_region_name_en}{UN region name, English.}
\item{un_subregion_name_en}{UN subregion name, English.}
\item{un_intermediate_region_name_en}{UN intermediate region name, English.}
\item{un_region_name_ru}{UN region name, Russian.}
\item{un_subregion_name_ru}{UN subregion name, Russian.}
\item{un_intermediate_region_name_ru}{UN intermediate region name, Russian.}
\item{un_region_name_fr}{UN region name, French.}
\item{un_subregion_name_fr}{UN subregion name, French.}
\item{un_intermediate_region_name_fr}{UN intermediate region name, French.}
\item{un_region_name_es}{UN region name, Spanish.}
\item{un_subregion_name_es}{UN subregion name, Spanish.}
\item{un_intermediate_region_name_es}{UN intermediate region name, Spanish.}
\item{un_region_name_ar}{UN region name, Arabic.}
\item{un_subregion_name_ar}{UN subregion name, Arabic.}
\item{un_intermediate_region_name_ar}{UN intermediate region name, Arabic.}
\item{un_region_name_zh}{UN region name, Chinese.}
\item{un_subregion_name_zh}{UN subregion name, Chinese.}
\item{un_intermediate_region_name_zh}{UN intermediate region name, Chinese.}
\item{un_desa_region}{UN DESA region code.}
\item{un_desa_region_name_en}{UN DESA region name, English.}
\item{un_desa_subregion}{UN DESA subregion code.}
\item{un_desa_subregion_name_en}{UN DESA subregion name, English.}
\item{sdg_region}{SDG region code.}
\item{sdg_region_name_en}{SDG region name, English.}
\item{sdg_subregion}{SDG subregion code.}
\item{sdg_subregion_name_en}{SDG subregion name, English.}
\item{gbd_region}{IHME GBD region code.}
\item{gbd_region_name_en}{IHME GBD region name, English.}
\item{gbd_subregion}{IHME GBD subregion code.}
\item{gbd_subregion_name_en}{IHME GBD subregion name, English.}
\item{wb_region}{WB geographical region. English name as no code provided by WB.}
\item{wb_region_name_en}{WB geographical region name, English.}
\item{wb_ig_1987}{World Bank Income Group code, 1987 (FY 1989) classification.}
\item{wb_ig_1988}{World Bank Income Group code, 1988 (FY 1990) classification.}
\item{wb_ig_1989}{World Bank Income Group code, 1989 (FY 1991) classification.}
\item{wb_ig_1990}{World Bank Income Group code, 1990 (FY 1992) classification.}
\item{wb_ig_1991}{World Bank Income Group code, 1991 (FY 1993) classification.}
\item{wb_ig_1992}{World Bank Income Group code, 1992 (FY 1994) classification.}
\item{wb_ig_1993}{World Bank Income Group code, 1993 (FY 1995) classification.}
\item{wb_ig_1994}{World Bank Income Group code, 1994 (FY 1996) classification.}
\item{wb_ig_1995}{World Bank Income Group code, 1995 (FY 1997) classification.}
\item{wb_ig_1996}{World Bank Income Group code, 1996 (FY 1998) classification.}
\item{wb_ig_1997}{World Bank Income Group code, 1997 (FY 1999) classification.}
\item{wb_ig_1998}{World Bank Income Group code, 1998 (FY 2000) classification.}
\item{wb_ig_1999}{World Bank Income Group code, 1999 (FY 2001) classification.}
\item{wb_ig_2000}{World Bank Income Group code, 2000 (FY 2002) classification.}
\item{wb_ig_2001}{World Bank Income Group code, 2001 (FY 2003) classification.}
\item{wb_ig_2002}{World Bank Income Group code, 2002 (FY 2004) classification.}
\item{wb_ig_2003}{World Bank Income Group code, 2003 (FY 2005) classification.}
\item{wb_ig_2004}{World Bank Income Group code, 2004 (FY 2006) classification.}
\item{wb_ig_2005}{World Bank Income Group code, 2005 (FY 2007) classification.}
\item{wb_ig_2006}{World Bank Income Group code, 2006 (FY 2008) classification.}
\item{wb_ig_2007}{World Bank Income Group code, 2007 (FY 2009) classification.}
\item{wb_ig_2008}{World Bank Income Group code, 2008 (FY 2010) classification.}
\item{wb_ig_2009}{World Bank Income Group code, 2009 (FY 2011) classification.}
\item{wb_ig_2010}{World Bank Income Group code, 2010 (FY 2012) classification.}
\item{wb_ig_2011}{World Bank Income Group code, 2011 (FY 2013) classification.}
\item{wb_ig_2012}{World Bank Income Group code, 2012 (FY 2014) classification.}
\item{wb_ig_2013}{World Bank Income Group code, 2013 (FY 2015) classification.}
\item{wb_ig_2014}{World Bank Income Group code, 2014 (FY 2016) classification.}
\item{wb_ig_2015}{World Bank Income Group code, 2015 (FY 2017) classification.}
\item{wb_ig_2016}{World Bank Income Group code, 2016 (FY 2018) classification.}
\item{wb_ig_2017}{World Bank Income Group code, 2017 (FY 2019) classification.}
\item{wb_ig_2018}{World Bank Income Group code, 2018 (FY 2020) classification.}
\item{wb_ig_2019}{World Bank Income Group code, 2019 (FY 2021) classification.}
\item{wb_ig_2020}{World Bank Income Group code, 2020 (FY 2022) classification.}
\item{wb_ig_2021}{World Bank Income Group code, 2021 (FY 2023) classification.}
\item{wb_ig_2022}{World Bank Income Group code, 2022 (FY 2024) classification.}
\item{wb_ig_2023}{World Bank Income Group code, 2023 (FY 2025) classification.}
\item{wb_ig_2024}{World Bank Income Group code, 2024 (FY 2026) classification.}
}
}
\source{
\href{https://portal-uat.who.int/xmart4/REFMART}{World Health Organization country reference table, accessed July 4 2022.}

\href{https://unstats.un.org/unsd/methodology/m49/overview/}{United Nations Statistics Division, Country or Area Codes for Statistical Use tables; accessed August 15 2020.}

\href{https://datahelpdesk.worldbank.org/knowledgebase/articles/906519-world-bank-country-and-lending-groups}{World Bank income groups, fiscal years 1989 - 2021, and regions, accessed August 23 2022.}

\href{https://www.oecd.org/about/members-and-partners/}{Organisation for Economic Co-operation and Development membership list, accessed 1 February 2021.}

\href{http://ghdx.healthdata.org/gbd-2019}{Institute for Health Metrics and Evaluation Global Burden of Disease 2019 Socio-Demographic Index data, accessed June 24 2022.}

\href{https://www.un.org/development/desa/dpad/wp-content/uploads/sites/45/publication/ldc_list.pdf}{United Nations Committee for Development Policy, list of least developed countries, accessed 16 February 2021.}

\href{https://www.un.org/development/desa/pd/sites/www.un.org.development.desa.pd/files/aggregates_correspondence_table_2020_1.xlsx}{United Nations Department of Economic and Social Affairs, Population Division (2020). International Migrant Stock 2020.}

\href{http://ghdx.healthdata.org/sites/default/files/ihme_query_tool/IHME_GBD_2019_CODEBOOK.zip}{Global Burden of Disease 2019 country and regional classifications, Institute for Health Metrics and Evaluation, accessed July 4 2022.}
}
\usage{
countries
}
\description{
A dataset containing names, codes, and other attributes
for countries included in World Health Organization,
United Nations and World Bank datasets.
}
\keyword{datasets}
