% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{country_code_types}
\alias{country_code_types}
\title{Country code types available}
\usage{
country_code_types()
}
\value{
Character vector of country code types.
}
\description{
\code{country_code_types()} provides a character vector of all country code types
available in the \code{countries} data frame.
}
\examples{
country_code_types()

}
