% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gbd.R
\name{get_gbd_direct}
\alias{get_gbd_direct}
\alias{format_gbd_direct}
\title{Get/Format Institute for Health Metrics and Evaluation (IHME) Global
Burden of Disease (GBD) location hierarchy information}
\usage{
get_gbd_direct(
  gbd_year = "2021",
  url = c(`2021` =
    "http://ghdx.healthdata.org/sites/default/files/ihme_query_tool/IHME_GBD_2021_CODEBOOK.zip",
    `2019` =
    "http://ghdx.healthdata.org/sites/default/files/ihme_query_tool/IHME_GBD_2019_CODEBOOK.zip"),
  zip_fname = c(`2021` = "IHME_GBD_2021_HIERARCHIES_Y2024M05D16.XLSX", `2019` =
    "IHME_GBD_2019_GBD_LOCATION_HIERARCHY_Y2022M06D29.XLSX")
)

format_gbd_direct(
  dat,
  map_iso3_manual = dplyr::tibble(`Location Name` = c("Taiwan (Province of China)",
    "Netherlands", "C\\\\u00f4te d\\\\'Ivoire"), iso3 = c("TWN", "NLD", "CIV"))
)
}
\arguments{
\item{gbd_year}{[\code{character(1)}]\cr
String specifying the GBD year/version to get}

\item{url}{[\code{character()}]\cr
Named vector specifying GBD year/version specific urls to download zip
files from}

\item{zip_fname}{[\code{character()}]\cr
Named vector specifying GBD year/version specific file name to read from
the downloaded zip file}

\item{dat}{[\code{data.frame(1)}]\cr
The original GBD location hierarchy to be formatted to match the
\code{whoville::countries} object formatting}

\item{map_iso3_manual}{[\code{data.frame()}]\cr
Mapping between \verb{Location Name} and \code{iso3} for GBD location names that do
not exactly match one of the \code{whoville::countries} name columns.}
}
\value{
\verb{get_gbd()_direct} returns a [\code{tibble()}] with the location hierarchy
from the specified zip/xlsx file. \verb{format_gbd()_direct} returns a [\code{tibble()}]
with a row for each GBD level 3 location and the mapped \code{iso3} code.
}
\description{
Get/Format Institute for Health Metrics and Evaluation (IHME) Global
Burden of Disease (GBD) location hierarchy information
}
\examples{
\dontrun{
dat_gbd <- get_gbd_direct("2021") \%>\%
  format_gbd_direct()
}

}
