% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/member_states.R
\name{is_gbd_high_income}
\alias{is_gbd_high_income}
\title{Check GBD high-income classification status from ISO3 codes.}
\usage{
is_gbd_high_income(iso3)
}
\arguments{
\item{iso3}{Character vector of ISO3 codes.}
}
\value{
Logical vector.
}
\description{
\code{is_gbd_high_income()} takes in a vector of ISO3 codes and returns a logical vector
on whether that country is classified as high-income in the 2019 Global Burden
of Disease released by the Institute for Health Metrics and Evaluation.
}
