% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/member_states.R
\name{is_oecd_member}
\alias{is_oecd_member}
\title{Get OECD membership status from ISO3 codes.}
\usage{
is_oecd_member(iso3)
}
\arguments{
\item{iso3}{Character vector of ISO3 codes.}
}
\value{
Logical vector.
}
\description{
\code{is_oecd_member()} takes in a vector of ISO3 codes and returns a logical vector
on whether that country is an OECD member state or not.
}
