% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/member_states.R
\name{is_un_lldc}
\alias{is_un_lldc}
\title{Check land-locked developing country classification status from ISO3 codes.}
\usage{
is_un_lldc(iso3)
}
\arguments{
\item{iso3}{Character vector of ISO3 codes.}
}
\value{
Logical vector.
}
\description{
\code{is_un_lldc()} takes in a vector of ISO3 codes and returns a logical vector
on whether that country is classified as a land-locked developing country by the
United Nations.
}
