% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{is_who_member}
\alias{is_who_member}
\title{Get membership status from ISO3 codes.}
\usage{
is_who_member(iso3)
}
\arguments{
\item{iso3}{Character vector of ISO3 codes.}
}
\value{
Logical vector.
}
\description{
\code{is_who_member()} takes in a vector of ISO3 codes and returns a logical vector
on whether that country is a WHO member state or not.
}
