% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/member_states.R
\name{is_who_member_large}
\alias{is_who_member_large}
\title{Get large state membership status from ISO3 codes.}
\usage{
is_who_member_large(iso3)
}
\arguments{
\item{iso3}{Character vector of ISO3 codes.}
}
\value{
Logical vector.
}
\description{
\code{is_who_member_large()} takes in a vector of ISO3 codes and returns a logical vector
on whether that country is a large WHO member state or not. In this instance,
large member states are defined as member states with a total population >
90,000 in mid-year 2018, according to the World Population Prospects.
}
