% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{iso3_to_codes}
\alias{iso3_to_codes}
\title{Get country codes from ISO3 country codes.}
\usage{
iso3_to_codes(iso3, type)
}
\arguments{
\item{iso3}{Character vector of ISO3 codes.}

\item{type}{A character value specifying the type of country code to return.
All possible values available through \code{country_code_types()}.}
}
\value{
Character vector.
}
\description{
\code{iso3_to_codes()} takes in a vector of ISO3 codes and returns a country code
vector specified by the user.
}
