% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{iso3_to_regions}
\alias{iso3_to_regions}
\title{Get regions from ISO3 country codes.}
\usage{
iso3_to_regions(
  iso3,
  region = c("who_region", "un_region", "un_subregion", "un_intermediate_region",
    "sdg_region", "sdg_subregion", "un_desa_region", "un_desa_subregion", "wb_region",
    "wb_ig"),
  year = max(wb_ig_years()),
  name = FALSE,
  language = c("en", "es", "ru", "ar", "zh", "fr")
)
}
\arguments{
\item{iso3}{Character vector of ISO3 codes.}

\item{region}{Type of region to return. Can be any of "who_region" (default),
"un_region", "un_subregion", "wb_region", or "wb_ig".}

\item{year}{For World Bank Income Group only, specify the year to return.}

\item{name}{For UN region and subregion and WB region only. If \code{TRUE},
return official UN name instead of M49 code.}

\item{language}{For UN region and subregion only, if returning name.
A character value specifying the language of the country names.
Should be specified using the ISO2 language code.}
}
\value{
Character vector.
}
\description{
\code{iso3_to_regions()} takes in a vector of ISO3 codes and returns a vector of
specified regions. Currently, this can be codes or English names for
UN regions/subregions/intermediate regions, UN DESA regions/subregions,
SDG regions/subregions, WHO regions, and World Bank geographical regions and
Income Groups.
}
