% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/codes.R
\name{iso3_to_regions}
\alias{iso3_to_regions}
\title{Get regions from ISO3 country codes.}
\usage{
iso3_to_regions(iso3, region = "who_region", return = "code")
}
\arguments{
\item{iso3}{Character vector of ISO3 codes.}

\item{region}{Type of region to return. Can be
any of "who_region" (default), "un_region", "un_subregion", or "wb_ig".}

\item{return}{Either "code" (default) or language code to specify the region
name. Only "en" currently available.}
}
\value{
Character vector.
}
\description{
\code{iso3_to_regions()} takes in a vector of ISO3 codes and returns regions
specified by the user. Currently, this can be codes or English names for
UN regions/subregions, WHO regions, and World Bank Income Groups.
}
