% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_countries.R
\name{make_countries}
\alias{make_countries}
\title{Make the \code{whoville::countries} dataset}
\usage{
make_countries(
  wb_use_xmart = TRUE,
  unsd_special_use_xmart = TRUE,
  sdg_use_xmart = TRUE,
  include_undesa = FALSE,
  include_gbd = FALSE,
  include_oecd = FALSE
)
}
\arguments{
\item{wb_use_xmart}{[\code{logical(1)}]\cr
Whether to get the world bank income groups and regions from xMart or
directly from the world bank website. Default is TRUE.}

\item{unsd_special_use_xmart}{[\code{logical(1)}]\cr
Whether to get the special unsd classifications (ldc, lldc, sids) from xMart
or directly from the unsd website. Default is TRUE.}

\item{sdg_use_xmart}{[\code{logical(1)}]\cr
Whether to get the sdg regions from xMart or directly from the undesa wpp
files. Default is TRUE.}

\item{include_undesa}{[\code{logical(1)}]\cr
Whether to include the undesa regions when making the countries object.
Default is FALSE to enable quicker and less buggy creation of the object
for when users are attempting to make their own updated object.}

\item{include_gbd}{[\code{logical(1)}]\cr
Whether to include the gbd regions when making the countries object.
Default is FALSE to enable quicker and less buggy creation of the object
for when users are attempting to make their own updated object.}

\item{include_oecd}{[\code{logical(1)}]\cr
Whether to include oecd membership when making the countries object.
Default is FALSE to enable quicker and less buggy creation of the object
for when users are attempting to make their own updated object.}
}
\value{
[\code{tibble()}] with the updated data for the \code{whoville::countries}
object.
}
\description{
Make the \code{whoville::countries} dataset
}
\details{
\itemize{
\item Basic location codes (m49, iso3, iso2, etc)
\item WHO regions and location name translations
\item World Bank income groups
\item World Bank regions
\item United Nations Statistics Division (UNSD) m49 regions
and location name translations
\item United Nations Department of Economic and Social Affairs (UN DESA) and the
Sustainable Development Goals (SDG) regions
\item Institute for Health Metrics and Evaluation (IHME) Global Burden of
Disease (GBD) regions
\item Organisation for Economic Co-operation and Development (OECD) membership
}
}
\examples{
\dontrun{
countries_new <- make_countries()
}

}
