% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_countries.R
\name{make_countries}
\alias{make_countries}
\title{Make the \code{whoville::countries} dataset}
\usage{
make_countries()
}
\value{
[\code{tibble()}] with the updated data for the \code{whoville::countries}
object.
}
\description{
Make the \code{whoville::countries} dataset
}
\details{
\itemize{
\item Basic location codes (m49, iso3, iso2, etc)
\item WHO regions and location name translations
\item World Bank income groups
\item World Bank regions
\item United Nations Statistics Division (UNSD) m49 regions
and location name translations
\item United Nations Department of Economic and Social Affairs (UN DESA) and the
Sustainable Development Goals (SDG) regions
\item Institute for Health Metrics and Evaluation (IHME) Global Burden of
Disease (GBD) regions
\item Organisation for Economic Co-operation and Development (OECD) membership
\item Alternative and former location names
}
}
\examples{
\dontrun{
countries_new <- make_countries()
}

}
