% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/member_states.R
\name{un_ldcs}
\alias{un_ldcs}
\title{Get ISO3 codes for least-developed countries.}
\usage{
un_ldcs()
}
\value{
A character vector of ISO3 codes.
}
\description{
\code{un_ldcs()} returns ISO3 codes for countries classified as least
developed by the United Nations. Useful to expand data frames to explicitly
include missing data for countries.
}
