% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/member_states.R
\name{un_lldcs}
\alias{un_lldcs}
\title{Get ISO3 codes for land-locked developing countries.}
\usage{
un_lldcs()
}
\value{
A character vector of ISO3 codes.
}
\description{
\code{un_lldcs()} returns ISO3 codes for countries classified as land-locked
developing countries by the United Nations. Useful to expand data frames to explicitly
include missing data for countries.
}
