% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_un_m49.R
\name{get_un_m49_direct}
\alias{get_un_m49_direct}
\alias{format_un_m49_direct}
\alias{format_unsd_special_xmart}
\title{Get/Format the United Nations Statistics Division (UNSD) m49 regions
and location name translations.}
\usage{
get_un_m49_direct(
  url = "https://unstats.un.org/unsd/methodology/m49/overview/"
)

format_un_m49_direct(
  dat_unsd_m49,
  language_codes = c(ru = "russian", fr = "french", es = "spanish", ar = "arabic", zh =
    "chinese")
)

format_unsd_special_xmart(dat_ref_groups_current, unsd_m49s)
}
\arguments{
\item{url}{[\code{character(1)}]\cr
Url to scrape the UNSD table (in html format) from.}

\item{dat_unsd_m49}{[\code{data.frame(1)}]\cr
The returned object from \code{get_un_m49} to be formatted to match the
\code{whoville::countries} object formatting.}

\item{language_codes}{[\code{character()}]\cr
Named character vector with the values corresponding to the language names
used in \code{dat_unsd_m49} and names corresponding to the language code to use
in the returned formatted data frame.}

\item{dat_ref_groups_current}{[\code{data.frame(1)}]\cr
The xMart4 REFMART 'REF_GROUPS_CURRENT' table to be formatted to match the
\code{whoville::countries} object formatting.}

\item{unsd_m49s}{[\code{data.frame(1)}]\cr
The data returned from \code{format_un_m49} which is used to determine the complete
list of locations that unsd classifies as ldc, lldc or sids.}
}
\value{
\code{format_unsd_special_xmart} returns a [\code{tibble()}] with the UNSD special
groups for 'Least Developed Countries (LDCs)', 'Landlocked developing
countries (LLDCs)', and 'Small island developing States (SIDS)' formatted to
match the \code{whoville::countries} object.

\code{get_un_m49_direct} returns a named list of [\code{data.frame()}]s with the list names
corresponding to the labeled language. \code{format_un_m49_direct} returns a [\code{tibble()}]
with the same data formatted to match the \code{whoville::countries} object.
}
\description{
Get/Format the United Nations Statistics Division (UNSD) m49 regions
and location name translations.
}
\examples{
\dontrun{
dat_unsd_special <- format_unsd_special_xmart(
  dat_ref_groups_current =  get_who_public_xmart(
    url = "https://xmart-api-public.who.int/REFMART/REF_GROUPS_CURRENT"
  ),
  dat_ref_country = get_who_public_xmart(
    url = "https://xmart-api-public.who.int/REFMART/REF_COUNTRY"
  ) \%>\%
    format_who_xmart_ref_country()
)

un_m49 <- get_un_m49_direct() \%>\%
  format_un_m49_direct()
}

}
