% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/member_states.R
\name{un_sids}
\alias{un_sids}
\title{Get ISO3 codes for small island developing states.}
\usage{
un_sids()
}
\value{
A character vector of ISO3 codes.
}
\description{
\code{un_sids()} returns ISO3 codes for countries classified as small island
developing states by the United Nations. Useful to expand data frames to explicitly
include missing data for countries.
}
