% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_wb.R
\name{get_wb_ig_direct}
\alias{get_wb_ig_direct}
\alias{get_wb_reg_direct}
\alias{format_wb_ig_direct}
\alias{format_wb_reg_direct}
\alias{format_wb_ig_xmart}
\alias{format_wb_reg_xmart}
\title{Get/Format the World Bank Income Group and Region country classifications.}
\usage{
get_wb_ig_direct(
  url = "https://ddh-openapi.worldbank.org/resources/DR0095334/download",
  sheet = "Country Analytical History",
  skip_years = 5,
  skip_data = 10
)

get_wb_reg_direct(
  url = "https://ddh-openapi.worldbank.org/resources/DR0095333/download",
  sheet = "List of economies"
)

format_wb_ig_direct(dat_raw)

format_wb_reg_direct(dat_raw)

format_wb_ig_xmart(dat_raw)

format_wb_reg_xmart(dat_raw)
}
\arguments{
\item{url}{[\code{character(1)}]\cr
Url to download the country classifications directly from the WB website.}

\item{sheet}{[\code{character(1)}]\cr
The name of the sheet within the xlsx document where the country
classifications are stored.}

\item{skip_years}{[\code{integer(1)}]\cr
The number of rows to skip when reading the XLSX document before reaching
the row listing the years available in the historical classification by income group document.}

\item{skip_data}{[\code{integer(1)}]\cr
The number of rows to skip when reading the XLSX document before reaching
the data rows with the country classifications.}

\item{dat_raw}{[\code{data.frame(1)}]\cr
The original country classification data from the World Bank or xMart4
REFMART tables to be formatted to match the \code{whoville::countries} object
formatting.}
}
\value{
\code{get_wb_ig_xmart} & \code{get_wb_reg_xmart} return a [\code{tibble()}] with data
formatted to match the \code{whoville::countries} object formatting.

\code{get_wb_ig_direct} and \code{get_wb_reg_direct} return a [\code{tibble()}] with data
from the specified World Bank XLSX document. \code{format_wb_ig_direct} and
\code{format_wb_reg_direct} return a [\code{tibble()}] with the same data formatted
to match the \code{whoville::countries} object formatting.
}
\description{
Each fiscal year, the World Bank updates their income group country
classifications and releases the files
\href{https://datahelpdesk.worldbank.org/knowledgebase/articles/906519-world-bank-country-and-lending-groups}{here}.

The links to the "historical classification by income in XLSX format" and
the "current classification by income in XLSX format" include the World Bank
Income Group and World Bank Region classifications respectively. The
\code{get_wb_ig_direct} & \code{get_wb_reg_direct} download these files and read in the
corresponding sheet from the XLSX document. \code{format_wb_ig_direct} and
\code{format_wb_reg_direct} then format the data to match the \code{whoville::countries}
object formatting.

Alternatively (preferably), this same data is extracted and uploaded to the
xMart4 REFMART 'REF_GROUPS_CURRENT' & 'REF_GROUPS_FULL' tables.
\code{format_wb_ig_xmart} & \code{format_wb_ref_xmart} format data from these tables
to match the \code{whoville::countries} object formatting.
}
\examples{
\dontrun{
wb_ig_xmart <- get_who_public_xmart(
  url = "https://xmart-api-public.who.int/REFMART/REF_GROUPS_FULL?$filter=GROUP_TYPE_CODE eq 'WB_INCOME'"
) \%>\%
  format_wb_ig_xmart()
wb_ig_direct <- get_wb_ig_direct() \%>\%
  format_wb_ig_direct()

wb_reg_xmart <- get_who_public_xmart(
  url = "https://xmart-api-public.who.int/REFMART/REF_GROUPS_CURRENT?$filter=GROUP_TYPE_CODE eq 'WB_REGION'"
) \%>\%
  format_wb_reg_xmart()
wb_reg_direct <- get_wb_reg_direct() \%>\%
  format_wb_reg_direct()
}

}
