% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wb_ig_years}
\alias{wb_ig_years}
\title{Years of WB IG classifications available}
\usage{
wb_ig_years()
}
\description{
\code{wb_ig_years()} provides a numeric vector of all years of World Bank Income
Group classifications available in the \code{countries} data frame.
}
\examples{
wb_ig_years()

}
