% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/member_states.R
\name{who_member_states}
\alias{who_member_states}
\title{Get ISO3 codes for WHO member states.}
\usage{
who_member_states(include = c("all", "small", "large"))
}
\arguments{
\item{include}{Subset of member states to return.
* "all": All member states (the default)
* "small": Small member states (2018 population < 90,000)
* "large": Large member states (2018 population >= 90,000)}
}
\value{
A character vector of ISO3 codes.
}
\description{
\code{who_member_states()} returns ISO3 codes for WHO member states. By default,
it returns all countries, but can be subset to only small (less than 90,000
population) or large states. Useful to expand data frames to explicitly include
missing data for countries.
}
