% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whoville-package.R
\docType{package}
\name{whoville-package}
\alias{whoville}
\alias{whoville-package}
\title{whoville: Country Codes, Names, and Other Metadata from the WHO, UN and WB}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

Provides a collated dataset with official country names and codes from the World Health Organization and United Nations, alongside additional metadata and utility functions to easily pull information into existing data frames.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/gpw13/ghost}
}

}
\author{
\strong{Maintainer}: Seth Caldwell \email{caldwellst@gmail.com}

}
\keyword{internal}
